sub-94: Refactored Lambda functions, added tests, and updated deployment scripts
