--- Dockerfile
@@ -1,12 +1,0 @@
-# Use the official AWS Lambda Python 3.9 base image
-FROM public.ecr.aws/lambda/python:3.9
-
-COPY requirements.txt ${LAMBDA_TASK_ROOT}
-
-RUN yum -y install git
-RUN pip install -r requirements.txt
-
-COPY lambda_function.py ${LAMBDA_TASK_ROOT}
-COPY config.json ${LAMBDA_TASK_ROOT}
-
-CMD [ "lambda_function.lambda_handler" ]
